bugfix/CMS-158 clean Architecture uyumu sağlandı, PII verileri temizlendi ve Site entity audit hataları giderildi

+++ backend/src/main/java/com/backend/domain/repository/ComponentRepository.java
@@ -47,5 +47,5 @@ public interface ComponentRepository {
 
     long countByStatus(ComponentStatus status);
 
-    int countByCreatedAtAfter(LocalDateTime date);
+    long countByCreatedAtAfter(LocalDateTime date);
 }
